tests sur les builtins exit, unset et export
